% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected.R
\name{get_selected}
\alias{get_selected}
\alias{get_selected.regsubsets}
\alias{get_selected.gbm}
\alias{get_selected.glmnet}
\title{Extract selected model terms from a (regularized) model}
\usage{
get_selected(model, ...)

\method{get_selected}{regsubsets}(model, ...)

\method{get_selected}{gbm}(model, ...)

\method{get_selected}{glmnet}(model, ...)
}
\arguments{
\item{model}{a fitted model object for which \code{\link{refit}} and
\code{\link{get_selected}} methods are defined}

\item{...}{other arguments to pass on to methods}
}
\value{
a logical matrix that summarizes
which covariates (=columns) were in the model in each iteration or for each
value of the regularization parameter (=rows). First row is the null model
(typically all FALSE).
}
\description{
Extract selected model terms from a (regularized) model
}
\details{
\code{get_selected.gbm} also counts variables used (only) for surrogate
splits as selected.
}
\section{Methods (by class)}{
\itemize{
\item \code{regsubsets}: method for \code{\link[leaps]{regsubsets}}

\item \code{gbm}: method for \code{\link[gbm]{gbm}} objects

\item \code{glmnet}: method for \code{\link[glmnet]{glmnet}} objects
}}

