% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability_paths.R
\name{get_stability_paths}
\alias{get_stability_paths}
\title{Extract stability paths}
\usage{
get_stability_paths(model, data, reps = 100, method = "subsample",
  strata = NULL, fraction = 0.5)
}
\arguments{
\item{model}{a fitted model object for which \code{\link{refit}} and
\code{\link{get_selected}} methods are defined}

\item{data}{the data used for the model fit, a \code{data.frame} or a list
containing \code{data.frame}s, vectors and/or matrices. See Details.}

\item{reps}{how many refits of \code{model} on resampled \code{data} to
perform}

\item{method}{resampling method, either "subsample" (without replacement) or
"bootstrap" (with replacement)}

\item{strata}{for subsampling, a vector with length nrow(data) defining the
strata}

\item{fraction}{subsampling fraction}
}
\value{
a <max. subsetsize + 1> x <covariates> matrix of relative selection
  frequencies with class \code{stabpath}. First row is the null model (no
  covariates, i.e., all 0s)
}
\description{
Extract stability paths from refits of a model object on resampled or
subsampled data.
}
\details{
List-shaped \code{data} arguments are necessary a) for models that use
weights or offsets and/or b) for modeling functions that receive the design
matrix and the response vector as arguments directly instead of a model
formula and a corresponding dataset. In these cases \code{data} has to be a
named list with names corresponding to the modelling function's respective
arguments (for example, \code{list(x = <design matrix>, y = <response>,
weights = <model weights>, offset = <model offsets>)} or
\code{list(data = <dataframe>, weights = <model weights>)})
}
