% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stabpath}
\alias{plot.stabpath}
\title{Plot stability paths}
\usage{
\method{plot}{stabpath}(x, y = NULL, label_position = c("rank", "mean"),
  path_pars = list(), label_pars = list(), ...)
}
\arguments{
\item{x}{a \code{stability_path}-object}

\item{y}{not used}

\item{label_position}{where to put the covariate labels: use rank or mean
inclusion probability (scaled to [0,1])}

\item{path_pars}{named list with optional arguments for \code{\link[graphics]{matplot}}-plotting of the stabiltiy path lines.
See source code for defaults.}

\item{label_pars}{named list with optional arguments for covariate labels via \code{\link[graphics]{mtext}}.
See source code for defaults.}

\item{...}{optional additional arguments handed over to \code{par}.}
}
\value{
invisibly returns the plotted object
}
\description{
Plot stability paths
}
