% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.R
\name{refit}
\alias{refit}
\alias{refit.regsubsets}
\alias{refit.gbm}
\alias{refit.glmnet}
\title{Refit a model on a new dataset}
\usage{
refit(model, new_data, ...)

\method{refit}{regsubsets}(model, new_data, ...)

\method{refit}{gbm}(model, new_data, ...)

\method{refit}{glmnet}(model, new_data, ...)
}
\arguments{
\item{model}{a fitted model object for which \code{\link{refit}} and
\code{\link{get_selected}} methods are defined}

\item{new_data}{the new data}

\item{...}{further arguments handed over to methods}
}
\value{
the fitted model object for \code{newdata}
}
\description{
This is roughly synonymous to \code{\link[stats]{update}} with
\code{update(model, data = new_data)}.
}
\section{Methods (by class)}{
\itemize{
\item \code{regsubsets}: method for \code{\link[leaps]{regsubsets}}

\item \code{gbm}: method for \code{\link[gbm]{gbm}} objects.

\item \code{glmnet}: method for \code{\link[glmnet]{glmnet}} objects
}}

