% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{resample.data.frame}
\alias{resample.list}
\title{Resample or subsample data rows}
\usage{
resample(data, method = c("subsample", "bootstrap"), strata = NULL,
  fraction = 0.5, ...)

\method{resample}{data.frame}(data, method = c("subsample", "bootstrap"),
  strata = NULL, fraction = 0.5)

\method{resample}{list}(data, method = c("subsample", "bootstrap"),
  strata = NULL, fraction = 0.5)
}
\arguments{
\item{data}{the data used for the model fit, a \code{data.frame} or a list
containing \code{data.frame}s, vectors and/or matrices. See Details.}

\item{method}{resampling method, either "subsample" (without replacement) or
"bootstrap" (with replacement)}

\item{strata}{for subsampling, a vector with length nrow(data) defining the
strata}

\item{fraction}{subsampling fraction}

\item{...}{further arguments handed over to methods}
}
\value{
the subsampled or resampled \code{data}
}
\description{
Resample or subsample data rows from a \code{data.frame} or a list containing
\code{data.frame}s, vectors and/or matrices.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: method for \code{data.frames}

\item \code{list}: entries in \code{data} have to be numeric vectors,
matrices or data.frames that all have the same length or number of rows.
}}

